requirements.txt
fastapi==0.103.1
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic[email]
uvicorn==0.23.2
alembic==1.12.0
pytest==8.4.2
pytest-dotenv==0.5.2
pytest-asyncio==1.2.0
pytest-aio==1.9.0
httpx==0.28.1
sqlalchemy==2.0.20
asyncpg
psycopg==3.1.10
psycopg-binary==3.1.12
fastapi-cache2[redis]
aiohttp
redis
aioredis